requirements.txt
langchain==0.3.7
langchain-openai==0.2.8
langchain-community==0.3.7
fastapi==0.115.5
uvicorn[standard]==0.32.1
pydantic==2.9.2
python-dotenv==1.0.1
requests==2.32.3
